refactor : code base reduced to multiple files
